+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ポップアップのレイアウトを調整してWindows環境での表示を改善しました。
 - リンク形式の未選択時に警告メッセージを表示するようにしました。
 
+### Changed
+
+- [Space]キーの押下でdescriptionの設定を変更できるようにしました。（キーバインドの変更）
+
 ## [1.0.3] - 2025-01-11
 
 ### Changed